struct(T12): split render modals into family modules
Add T12 to the structure handoff and pure-move render/modals.rs into
modals/{deals,charts,marketing,file,pickers} with a thin re-export mod.
Leaf draw fns are pub(crate) so nested re-exports stay visible to
render::draw. Parallel-safe with T7 (UI-only owns).
